[fix] Fixes issues where script stopped working again.
(╯°□°）╯︵ ┻━┻
For some reason, the commit worked last night but all of a sudden, it's not
working again.

Gosh I hate to work with JavaScript. Oh well, here's the buckload of changes on
this commit:

chore: Remove webpack-dev-server as this is no longer used to test
chore: Update webpack-cli to 4.2.0 after downgrading it to 3 in order for
       webpack-dev-server to work. But now it is removed and we are free to
       update it on newer one.
chore: events.js is now events.json.
chore: events is no longer included in support object. Now events are just imported
       at the script that processes the events. Well, in short, it is just an
       optimization.
fix:   This commit updates where script stopped working automatically for no
       reason. Though previous commit works after release but after it, it just gone.
       This fixes issues on regex on dates, changed up and tidied a little bit of
       wires and then it just works again. Expect again that this shit will stop
       working again soon. ヽ(✿ﾟ▽ﾟ)ノ

Frustrating but, I should return the table back again.
┳━┳ ノ( ゜-゜ノ)